# docs/api/openapi.yaml
...
openapi: 3.0.3
info:
  title: Hamlaneh API
  description: >
    Self-hosted team communication platform API. Phase 1.1 surface: health,
    authentication, current user, and the admin user-management endpoints that
    back the dashboard (public registration is off by default; users are
    created from the admin dashboard or via invite links). Phase 1.2 adds the
    messaging surface the chat shell needs: channels and 1:1 direct messages
    with membership, cursor-paged message history with idempotent send, edit
    and soft delete, per-user read positions, message search, and the
    WebSocket upgrade that carries realtime delivery. Phase 1.1b completes the
    identity surface: email password reset, TOTP two-step verification with
    recovery codes, and self-service session management.
  version: 0.6.0
  license:
    name: AGPL-3.0
    url: https://www.gnu.org/licenses/agpl-3.0.html
servers:
  - url: /
    description: Same-origin deployment behind Caddy.
...
paths:
...
  /api/v1/channels/{channelId}/files:
    post:
      operationId: uploadFile
      summary: Upload one file into a channel.
      description: >
        Backs the composer's paperclip. One file per request — the composer
        uploads several as several requests, each with its own progress and
        its own failure. The file is scoped to this channel from birth:
        membership is checked here, the later message references it by id,
        and a file never attached to a message is swept after 24 hours. That
        one rule — a file is readable exactly by the members of its channel —
        keeps the authorization story identical to every other channel
        resource.

        The declared content type is kept as the card's label, but the server
        trusts the bytes, not the label: a declared image/* whose bytes do
        not sniff as that image type is refused (415, content_type_mismatch),
        because images are the one kind served inline. Everything else is
        stored as an opaque blob and always served as a download —
        Content-Disposition attachment, nosniff, sandboxing CSP — so an
        uploaded SVG, HTML page or anything else that can carry script is a
        file the reader saves, never a page their browser runs. Images have
        their metadata stripped at ingest: EXIF carries GPS coordinates, and
        a photo shared in a chat must not quietly say where its author lives.
      security:
        - sessionCookie: []
      parameters:
        - $ref: "#/components/parameters/ChannelId"
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required: [file]
              properties:
                file:
                  type: string
                  format: binary
      responses:
        "201":
          description: >
            Stored. The attachment's url and thumbnail_url are signed and
            expiring — the files origin is cookie-less, so possession of a
            fresh URL is the credential, and every serialization of an
            Attachment mints fresh ones.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Attachment"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
        "403":
          $ref: "#/components/responses/Forbidden"
        "404":
          description: No such channel — including one the caller is not in.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "413":
          description: The file exceeds max_file_size_bytes (code file_too_large).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "415":
          description: >
            Declared image bytes that are not that image (code
            content_type_mismatch).
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Error"
        "429":
          $ref: "#/components/responses/RateLimited"
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: hamlaneh_session
      description: >
        HttpOnly+Secure+SameSite=Strict session cookie. Mutating requests must
        also send the X-Hamlaneh-CSRF header.
  parameters:
    ChannelId:
      name: channelId
      in: path
      required: true
      description: A channel or direct message the caller is a member of.
      schema: { type: string, format: uuid }
...
  responses:
    BadRequest:
      description: Request validation failed.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Missing, expired, or invalid session/credentials.
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Authenticated but not allowed (authz denied).
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    RateLimited:
      description: >
        Too many attempts; retry later. Always carries Retry-After, so a client
        can show a real countdown instead of guessing — without it a form has no
        honest way to say when the door reopens.
      headers:
        Retry-After:
          description: >
            Whole seconds until the caller's budget frees up. Present on every
            429 this server sends: one function writes them all and it sets the
            header unconditionally, so carrying it is a property of the code
            rather than something each handler has to remember.
          schema:
            type: integer
            minimum: 1
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
  schemas:
...
    Error:
      type: object
      required: [error]
      properties:
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Stable machine-readable code (e.g. invalid_credentials).
            message:
              type: string
              description: Human-readable English message; clients localize by code.
...
    Attachment:
      type: object
      description: >
        A file card in the message list. Read-only on Message: attachments are
        created by the Phase 1.3 upload pipeline, so until that slice lands the
        array is always empty. Every field here is one the design draws — name,
        type and size on the generic card; pixel dimensions and a thumbnail on
        the image card; a download control on both.
      required: [id, filename, content_type, size_bytes, url]
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
          maxLength: 255
        content_type:
          type: string
          description: The card's short type label ("PDF", "PNG") is derived from this.
        size_bytes:
          type: integer
          format: int64
          minimum: 0
        width:
          type: integer
          nullable: true
          description: Images only — the card's "1600 × 900" line.
        height:
          type: integer
          nullable: true
        url:
          type: string
          description: >
            Download target on the cookie-less files origin. Signed and
            expiring (about an hour): that origin carries no cookies, so a
            fresh URL is the credential, minted every time an Attachment is
            serialized for somebody entitled to see it. Stale ones answer 404
            — re-fetch the message for fresh links, never store these.
        thumbnail_url:
          type: string
          nullable: true
          description: >
            Images only — the card's preview, signed and expiring like url.
            Thumbnails are derivatives generated at ingest from the stripped
            original, bounded to 512px on the long edge.
...